+++ waspc/cli/src/Wasp/Cli/Command/Require/InWaspProject.hs
@@ -1,0 +1,53 @@
+module Wasp.Cli.Command.Require.InWaspProject
+  ( InWaspProject (InWaspProject),
+  )
+where
+
+import Control.Monad (unless, when)
+import Control.Monad.Error.Class (throwError)
+import Control.Monad.IO.Class (liftIO)
+import Data.Data (Typeable)
+import Data.Maybe (fromJust)
+import StrongPath (Abs, Dir, Path')
+import qualified StrongPath as SP
+import System.Directory (doesFileExist, doesPathExist, getCurrentDirectory)
+import qualified System.FilePath as FP
+import Wasp.Cli.Command (CommandError (CommandError), Requirable (checkRequirement))
+import Wasp.Project.Common (WaspProjectDir)
+import qualified Wasp.Project.Common as Project.Common
+
+-- | Require a Wasp project to exist near the current directory. Get the
+-- project directory by pattern matching on the result of 'Wasp.Cli.Command.require':
+--
+-- @
+-- do
+--   InWaspProject waspProjectDir <- require
+-- @
+newtype InWaspProject = InWaspProject (Path' Abs (Dir WaspProjectDir)) deriving (Typeable)
+
+instance Requirable InWaspProject where
+  checkRequirement = do
+    -- Recursively searches up from CWD until @.wasproot@ file is found, or
+    -- throw an error if it is never found.
+    currentDir <- fromJust . SP.parseAbsDir <$> liftIO getCurrentDirectory
+    findWaspProjectRoot currentDir
+    where
+      findWaspProjectRoot currentDir = do
+        let absCurrentDirFp = SP.fromAbsDir currentDir
+        doesCurrentDirExist <- liftIO $ doesPathExist absCurrentDirFp
+        unless doesCurrentDirExist (throwError notFoundError)
+        let dotWaspRootFilePath = absCurrentDirFp FP.</> SP.fromRelFile Project.Common.dotWaspRootFileInWaspProjectDir
+        isCurrentDirRoot <- liftIO $ doesFileExist dotWaspRootFilePath
+        if isCurrentDirRoot
+          then return $ InWaspProject $ SP.castDir currentDir
+          else do
+            let parentDir = SP.parent currentDir
+            when (parentDir == currentDir) (throwError notFoundError)
+            findWaspProjectRoot parentDir
+
+      notFoundError =
+        CommandError
+          "Wasp command failed"
+          ( "Couldn't find wasp project root - make sure"
+              ++ " you are running this command from a Wasp project."
+          )
