+++ waspc/cli/src/Wasp/Cli/Command/Telemetry/Project.hs
@@ -28,9 +28,9 @@ import qualified StrongPath as SP
 import qualified System.Environment as ENV
 import qualified System.IO as SIO
 import qualified System.Info
-import Wasp.Cli.Command (Command)
+import Wasp.Cli.Command (Command, require)
 import qualified Wasp.Cli.Command.Call as Command.Call
-import Wasp.Cli.Command.Require (InWaspProject (InWaspProject), require)
+import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
 import Wasp.Cli.Command.Telemetry.Common (TelemetryCacheDir)
 import Wasp.Cli.Command.Telemetry.User (UserSignature (..))
 import Wasp.Util (checkIfOnCi, ifM, isOlderThanNHours)
