+++ waspc/waspc.cabal
@@ -515,7 +515,11 @@ library cli-lib
     Wasp.Cli.Command.News.Fetching
     Wasp.Cli.Command.News.Listing
     Wasp.Cli.Command.News.LocalNewsState
-    Wasp.Cli.Command.Require
+    Wasp.Cli.Command.Require.DbConnectionEstablished
+    Wasp.Cli.Command.Require.GeneratedApp
+    Wasp.Cli.Command.Require.InWaspProject
+    Wasp.Cli.Command.Require.ValidNodeAndNpm
+    Wasp.Cli.Command.Require.WaspSpecAvailable
     Wasp.Cli.Command.Start
     Wasp.Cli.Command.Start.Db
     Wasp.Cli.Command.Studio
